+++ internal/consumer/options.go
@@ -38,6 +38,8 @@ type (
 		RebalanceDwellTime     time.Duration
 		MaxProcessingTime      time.Duration // amount of time a partitioned consumer will wait during a drain
 		ConsumerMode           cluster.ConsumerMode
+		ProducerMaxMessageByes int
+		FetchDefaultBytes      int32
 		OtherConsumerTopics    []Topic
 	}
 )
@@ -53,6 +55,8 @@ func DefaultOptions() *Options {
 		MaxProcessingTime:      250 * time.Millisecond,
 		OffsetPolicy:           sarama.OffsetOldest,
 		ConsumerMode:           cluster.ConsumerModePartitions,
+		FetchDefaultBytes:      10 * 1024 * 1024, // 10MB.
+		ProducerMaxMessageByes: 10 * 1024 * 1024, // 10MB
 		OtherConsumerTopics:    make([]Topic, 0, 10),
 	}
 }