+++ internal/consumer/clusterConsumer_test.go
@@ -68,14 +68,12 @@ func TestClusterConsumerTestSuite(t *testing.T) {
 func (s *ClusterConsumerTestSuite) SetupTest() {
 	topic := kafka.ConsumerTopic{
 		Topic: kafka.Topic{
-			Name:       "unit-test",
-			Cluster:    "production-cluster",
-			BrokerList: nil,
+			Name:    "unit-test",
+			Cluster: "production-cluster",
 		},
 		DLQ: kafka.Topic{
-			Name:       "unit-test-dlq",
-			Cluster:    "dlq-cluster",
-			BrokerList: nil,
+			Name:    "unit-test-dlq",
+			Cluster: "dlq-cluster",
 		},
 	}
 	s.topic = topic.Topic.Name