+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 v0.1.9 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Tune producer max message bytes to 10mb.
+- Tune consumer default fetch bytes to 10mb.
 
 
 v0.1.8 (2018-06-04)